💎 Add support for repos without a Gemfile.lock
We want to run `balto-rubocop` on a gem. It is best practice for gems to
not commit `Gemfile.lock`. So we want to fallback to a `project.gemspec`
file if it exists.

A `Gemfile.lock` has the benefit of having absolute versions determined.
This allows us to run a command of the form...

```bash
gem install -N "rubocop:0.77.0" "rubocop-performance:1.5.2"
```

...installing multiple gems at once.

When working with a `gemspec` file, we don't have absolute versions. We
might have a version with multiple requirements that are not yet
resolved, something like...

```ruby

gemspec.add_development_dependency "rubocop", "~> 0.50.0", "<= 60"
```

There's no way (that I could find) to represent multiple requirements,
and install multiple gems in one go. So for the gemspec strategy, we
install dependencies one at a time.