Use bundler/inline to install gems
We found a specific instance where an app needed to specify the
version of `rubocop` and `rubocop-ast`, but the default `gem install`
resulted in a much newer version of `rubocop-ast` that broke.

Here's an example of the issue:

```
❯ gem install --explain rubocop-ast:0.0.3 rubocop:0.84.0
Gems to install:
  ast-2.4.1
  parser-2.7.1.5
  rubocop-ast-0.0.3
Gems to install:
  parallel-1.19.2
  ast-2.4.1
  parser-2.7.1.5
  rainbow-3.0.0
  rexml-3.2.4
  strscan-1.0.3
  rubocop-ast-0.7.0
  ruby-progressbar-1.10.1
  unicode-display_width-1.7.0
  rubocop-0.84.0
```

That will automatically install the new version of `rubocop-ast` and
use it, even with the explicit version of `rubocop-ast`, and that's
failing. `bundler/inline` uses Bundler's resolution algorithm to
properly get us only what we're really asking for, and then helps us
make sure we only use the gems we wanted to install.