Setup local testing
A while back, I learned about [act][]: a tool that lets you run actions
locally. By default act uses a very stripped down docker container that
only runs node. That would be great for balto-eslint, but we need ruby
for this action.

Act allows us to define an alternate docker image. Unfortunately, it's
somewhere in the range of 6-12gb — but it has almost(?) everything that
the regular github workflow action has.

Anyway, testing locally is definitely optional, but I wanted to get it
up and running to get our feedback loop a bit smaller for some work I'm
planning on in the future.

[act]: https://github.com/nektos/act