Fix head_sha for pull_request events
When support was added for push events, we switched to ENV['GITHUB_SHA']
for the check run's head_ref. This is apparently correct for push events
but not for pull request events. Until I can test push events more
easily to find a common attribute we can use, this conditional should do
just fine.