Add option to install all gems
For large apps that have a bunch of dependencies, installing the fewest
gems required to run RuboCop can provide quite the time savings. For
smaller projects like gems, it's not as big of a deal. And especially in
the case of a gem that integrates with RuboCop, getting everything
installed just right is challenging.

In lieu of adding more smarts to the existing gem install strategies, I
opted for an escape hatch option that lets a user say they're fine with
the full bundle install. In this case, we'll also run `rubocop` with
`bundle exec` to make sure it's got access to everything in the bundle
that it just installed.