Add CreateMinimalGemfile strategy
TL;DR: New flow which hinges on either using an existing Gemfile.lock or
an existing Gemfile that we can use to create a Gemfile.lock to be
parsed for the minimal gems required for RuboCop. Build a new,
temporary, minimal Gemfile from there that can be used by other commands
to install those gems, run RuboCop, etc. in a predictable, repeatable
bundler environment.

One of the initial aims of balto-rubocop was to create an action that
could selectively install the minimal dependencies necessary to run
RuboCop for a given project. The initial use of bundler/inline worked
well for a while. We've also made affordances for the differences
between a Ruby project that has a Gemfile.lock (i.e. most apps) and
other projects like gems that don't generally have a Gemfile.lock.

We've covered most cases pretty well, but we're starting to see more an
more affordances that we need to make (e.g. not all gems put
dev dependencies in the gemspec, a larger set of non rubocop- gems,
version conflicts with default gems, etc.). I'd propose that keeping the
minimal gem install is worthwhile.

This strategry seeks to keep that minimal install but split things up
and use existing bundler tooling to build a setup that's more
deterministic and predictable: create a new gemfile with only the gems
required, and let all subsequent commands use this gemfile.

In the case of a project with a gemspec, that should always be sourced
in the project's Gemfile, which should be part of source control. But
calling `bundle lock` for projects like those (or any others that are
missing a Gemfile.lock), we'll make sure that we get all of the gems
used for development, and we can then use a single strategy for parsing
the lockfile for gem information.