Use CreateMinimalGemfile
In the interest of predictability and flexibility, we can now lean on
the new CreateMinimalGemfile strategy, which builds a new, minimal
Gemfile for running RuboCop.

The steps look like this:

1. Create a working directory for the files we'll create in this action
2. Create the minimal gemfile
3. Install gems using minimal gemfile
4. Run action using minimal gemfile
5. Clean up working directory

This change also catches any errors and sets a failure exit code.
Previously, errors at certain stages would show up in the action logs
but would still report a successful check run.