Document additionalGems input
Sometimes there are other gems that a project's RuboCop will depend on
that don't begin with "rubocop". We already called this out when we
added explicit support for `standard`. Instead of expanding that list
manually, let's add an input so users of this action can do that on
their own.

Since we already added `standard`, we can keep that as the default, at
least for now.

Also added documentation for the existing rootDirectory input