Don't use a directory for minimal gemfile
This was nice for having a place to do work and clean it up all at once,
but it causes problems with gems using a `path` option. Instead of
trying to keep relative paths straight, let's use a specially named
gemfile. Since we might be creating a Gemfile.lock, we can use git to
reset the working directory with any files we may have created, which
avoids us having to do double work or pass more info between JS and
Ruby.