Ensure we can bundle install
It's possible for previous steps in a workflow to set bundler to
deployment mode, which will prevent us from being able to `bundle
install` even with our custom gemfile.

The ruby/setup-ruby action uses `bundle config --local` to set the
deployment flag. Since Bundler reads from this first, we can't override
the setting with an ENV variable. We'd have to use a `bundle config`
call of our own. But then we'd really need to check the value before we
run things and then clean up after ourselves.

Instead of doing all of that work, we can change the BUNDLE_APP_CONFIG
directory to /dev/null, which effectively removes all local config
settings just for our commands.