Remove standard from default gems
A previous version of this action added standard as one of the gems to
look for. Now that we have the additionalGems input, which is more
flexible, projects needing standard can add it to that input.

With us moving towards a new major version, let's remove this from the
default for that input (which was just a bridge solution for backwards
compatibility).