chore: Update readme
Permissions
---
Because we're not making network requests for the checks, we don't need
special permissions any more.

Changing job name from lint -> rubocop
---
When we created the check run, we were able to specify the name
displayed on the annotation. Now the job title will be used, so let's
keep it close to identical.